Convert AcademicDatesBlock test from Kernel to Unit
The block is constructor-injected, so its build, validate, and submit
logic can be exercised with mocks alone. Removes the kernel bootstrap
the test never used.

Refs #9877
